requirements.txt
absl-py==0.2.2
backcall==0.1.0
bleach==2.1.3
bokeh==0.12.16
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
davisinteractive==0.1.9
decorator==4.3.0
distributed==1.21.8
entrypoints==0.2.3
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.3.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
locket==0.2.0
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
msgpack-python==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.5.0
numpy==1.14.3
olefile==0.45.1
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psutil==5.4.5
ptyprocess==0.5.2
pycocotools==2.0.0
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
sortedcontainers==1.5.10
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==1.0.0
torchvision==0.2.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
zict==0.1.3
